+++ app/src/main/java/com/harsh/samples/thisweektvshow/domain/model/TvShowSeason.kt
@@ -3,7 +3,7 @@ package com.harsh.samples.thisweektvshow.domain.model
 data class TvShowSeason(
     val id: Long,
     val name: String,
-    val noOfEpisodes: Byte,
+    val noOfEpisodes: Int,
     val seasonNumber: Byte,
     val voteAvg: Float
 )
